fix(search): Using wrong `toString` method for getting excerpt and heading (#4290)
We stringify heading and mdast to extract searchable text for algolia. But the `toString` method used is meant for html AST not mdx. So every mdx-specific nodes (like code nodes) was not included.

Using `toString` from `mdast-util-to-string` does includes those special nodes in the stringify process. But it also stringify the `jsx` nodes (ex: `import File from '...'`) so we need to extract only paragraph nodes to get the excerpt.

This should step-up the search experience a little bit.