fix(algolia): wrong key when adding custom table options (#5254)
When indexing mdx pages to algolia, we retrieve our custom `OptionsTable` and `ActionOptionsTable`, we extract the config schema definitions they render to pass it to algolia in order to provide searching through those tables.

Those tables have changed and takes a `def` props instead of a `name` prop to point to the definition they render, that was breaking things.